+++ src/main/java/capstone/is4103capstone/entities/finance/Invoice.java
@@ -5,6 +5,7 @@ import capstone.is4103capstone.entities.supplyChain.Vendor;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table
@@ -17,7 +18,7 @@ public class Invoice extends DBEntityTemplate {
 
     private String description;
 
-    private Double paymentAmount;
+    private BigDecimal paymentAmount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "vendor_id")
@@ -27,24 +28,24 @@ public class Invoice extends DBEntityTemplate {
     public Invoice() {
     }
 
-    public Invoice(String objectName, String code, String description, Double paymentAmount) {
+    public Invoice(String objectName, String code, String description, BigDecimal paymentAmount) {
         super(objectName, code);
         this.description = description;
         this.paymentAmount = paymentAmount;
     }
 
-    public Invoice(PurchaseOrder purchaseOrder, String description, Double paymentAmount, Vendor vendor) {
+    public Invoice(PurchaseOrder purchaseOrder, String description, BigDecimal paymentAmount, Vendor vendor) {
         this.purchaseOrder = purchaseOrder;
         this.description = description;
         this.paymentAmount = paymentAmount;
         this.vendor = vendor;
     }
 
-    public Double getPaymentAmount() {
+    public BigDecimal getPaymentAmount() {
         return paymentAmount;
     }
 
-    public void setPaymentAmount(Double paymentAmount) {
+    public void setPaymentAmount(BigDecimal paymentAmount) {
         this.paymentAmount = paymentAmount;
     }
