+++ src/main/java/capstone/is4103capstone/entities/supplyChain/ContractLine.java
@@ -4,19 +4,20 @@ import capstone.is4103capstone.configuration.DBEntityTemplate;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table
 public class ContractLine extends DBEntityTemplate{
     private String merchandiseCode;
-    private Double price;
+    private BigDecimal price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "contract")
     @JsonIgnore
     private Contract contract;
 
-    public ContractLine(String merchandiseCode, Double price) {
+    public ContractLine(String merchandiseCode, BigDecimal price) {
         this.merchandiseCode = merchandiseCode;
         this.price = price;
     }
@@ -32,11 +33,11 @@ public class ContractLine extends DBEntityTemplate{
         this.merchandiseCode = merchandiseCode;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
